Routing: Hook appearing/disappearing neighbors
In the last commits, the CLA handling was switched to the CLA Manager.
The manager now sends ConvergenceStatus messages to the Core, which
forwards them to its RoutingAlgorithm. This includes the appearance and
disappearance of neighboring nodes. Two further RoutingAlgorithm
interface methods have been added for this purpose.

Furthermore, this closes #11.