MTCP Client: TCP keepalive
Based on the strict unidirectional nature of MTCP, there is no keepalive
or timeout. If a link between two MTCP nodes disappears - e.g., the
cable is unplugged - there will be no FIN packet and it seems like the
connection is still active.

This commits introduces a keepalive to monitor the connection and detect
failures. However, this solution is still far from perfect and does not
work all the time. Thus, I would not recommend MTCP for rupture
environments at this time.