DtnEndpoint: require proper URL
The 25th draft of the Bundle Protocol 7 includes an Augmented
Backus-Naur Form (ABNF) to specify the grammar of the URI schemas. This
ABNF requires some dtn URI to be of the following form:

  dtn://NODE-NAME/OPTIONAL-VARIOUS-CHARS

Furthermore, the null endpoint "dtn:none" is specified next to the ABNF
as before.

Currently I am not quite sure how to handle this and posted about this
on the dtn mailing list[0]. However, now I am going to enforce this more
strict ABNF for all dtn URIs which are not the null endpoint.

This resulted in a small modification of the DtnEndpoint type and lots
of adjustments in all the tests.

[0]:https://mailarchive.ietf.org/arch/msg/dtn/12-NZl8tF3Xu_QUruWGzWMZM2LI/