DtnEndpoint: Change struct to match ABNF parts
Previously, a DtnEndpoint stored the scheme specific part (SSP)
representing the endpoint. This representation is kind of static and
mostly used within CBOR encoding on the wire.

This commit introduces a change to the fields of the struct to represent
the fields of the Augmented Backus-Naur Form (ABNF):
- a node name, and
- an optional demux part.

Furthermore, an explicit flag states if some DtnEndpoint is the null
endpoint "dtn:none".