TCPCL: Simplify stages by removing concurrency
Each stage was executed within its own Goroutine. On the one hand, this
is a nice encapsulation for testing or the like. However, on the other
hand, this is just overhead and introduces more synchronization issues.
Thus, stages are "reduced" to a blocking Handler function, which is used
within the StageHandler, which itself is concurrent.