Enable parsing Primary Blocks without a CRC
So far, the code only allowed parsing primary blocks with a CRC present.
However, only in ietf-draft-bpbis-14 the CRC was mandatory. Since
revision 15, either a CRC or an integrity block (BPSec) was required for
the primary block. Due to the missing implementation of BPSec, the code
required the presence of a CRC.

But this makes compatibility with the standard difficult.

Now parsing primary blocks without a CRC should work. When creating a
new primary block, an attempt is made to enforce a CRC, since this is
necessary in the absence of an integrity block. This should be changed
when BPSec has landed in dtn7-go.

This closes #34. Thanks for reporting!