Workflow for extended tests (#48)
* Workflow for extended tests

As can be seen in recent issues and PRs (#39, #40, #41, #42, #43, #45), there
is an issue in dtn7-go that results in a deadlock after some time, making it
impossible to sent bundles. To be able to test and debug this issue, some more
extensive tests are required. Thus, we developed a test infrastructure to
simulate multiple virtual nodes using the
[CORE emulator](https://github.com/coreemu/core).
This tests environment is located in the
[dtn7-playground](https://github.com/dtn7/dtn7-playground) repository. To be
able to trigger these tests, this PR adds a workflow file, that uses the
[repository_dispatch event](https://docs.github.com/en/actions/reference/events-that-trigger-workflows#repository_dispatch)
sent to the dtn7-playground repo.

* Fix copyright header

Co-authored-by: Artur Sterz <github@artursterz.de>