Harmonise `Bundle` and `BundleDescriptor` API
Both structs have an ID() method, however in the case of `Bundle`, this returned a `BundleID`-struct, while in case of `BundleDescriptor`, it returned a string-representation of the Id.

This caused confusion on multiple occasions, since in many cases theses two types are used in close proximity to one another and if you used the `Bundle`'s `ID`-method you might end up having the raw `EndpointID` struct dumped in you logs.

Now in both cases, the `ID`-method returns an `EndpointID` struct, and in any case where we want to log a bundle's id, we call the `String`-method explicitly.