Implement QUIC convergence layer ("quicl")
This should serve as a nice middle-ground between mtcp, which is simple but has very lmited functionality, and tcpcl which is overengineerd.

While this convergence layer should have basically all the features of tcpcl, most of the heavy-lifting is done by the QUIC implementation, so that we don't need to worry about session management, multiplexing, or encryption.